opengl: split into convenience library
Create a libvlc_opengl and libvlc_opengles library that are built only
if one other target is needing it, avoiding to compile the OpenGL code
once per module using it and removing the need for OPENGL_COMMON* vars.

As the fact we're using OpenGL or OpenGL ES is defined at compile time,
the clients must use the correct variant depending on what they use.